SE-21008: Maven Plugin: Failing to build application from command line (#501)
The artifact installer is assuming that every dependency has a pom file.
This is not the case for the tools.jar that is obtained from the jre.
This fix consists of checking that the pom exists before trying to copy
it to target folder.

+++ mule-packager/src/main/java/org/mule/tools/api/repository/ArtifactInstaller.java
@@ -98,7 +98,9 @@ public class ArtifactInstaller {
     if (!srcPomFile.exists()) {
       srcPomFile = new File(srcPomFolder, POM_FILE_NAME);
     }
-    copyFile(srcPomFile, destinationPomFile);
+    if (srcPomFile.exists()) {
+      copyFile(srcPomFile, destinationPomFile);
+    }
   }
 
   @Deprecated
